# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: openai
    description: AI chat operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /openai/conversations:
    post:
      operationId: createConversation
      tags: [openai]
      summary: Create a new conversation
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /openai/conversations/{id}/messages:
    post:
      operationId: sendMessage
      tags: [openai]
      summary: Send a message and stream AI response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of AI response
          content:
            text/event-stream:
              schema:
                type: string
    get:
      operationId: getMessages
      tags: [openai]
      summary: Get conversation messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Conversation:
      type: object
      properties:
        id:
          type: string
        createdAt:
          type: string
      required:
        - id
        - createdAt
    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content